Groups tweaks
Making adjustments for groups to accomplish the assignment.  Currently, the assumption is that when Product E is put into the group pool, it is accessible to Person Y even though it technically belongs to Person X.  Right now, a second group internal transaction would be needed to complete the barter.